move tests from test_scheduled_tasks to test_nightly_tasks
